+++ etna/transforms/outliers/base.py
@@ -6,6 +6,7 @@ from typing import Optional
 
 import numpy as np
 import pandas as pd
+from deprecated import deprecated
 
 from etna.datasets import TSDataset
 from etna.transforms.base import ReversibleTransform
@@ -26,9 +27,32 @@ class OutliersTransform(ReversibleTransform, ABC):
         """
         super().__init__(required_features=[in_column])
         self.in_column = in_column
-        self.outliers_timestamps: Optional[Dict[str, List[pd.Timestamp]]] = None
-        self.original_values: Optional[Dict[str, List[pd.Timestamp]]] = None
+
+        self.segment_outliers: Optional[Dict[str, pd.Series]] = None
+
         self._fit_segments: Optional[List[str]] = None
+
+    @property
+    @deprecated(
+        reason="Attribute `outliers_timestamps` is deprecated and will be removed! Use `segment_outliers` instead.",
+        version="3.0",
+    )
+    def outliers_timestamps(self) -> Optional[Dict[str, List[pd.Timestamp]]]:
+        """Backward compatibility property."""
+        if self.segment_outliers is not None:
+            return {segment: outliers.index.to_list() for segment, outliers in self.segment_outliers.items()}
+        return None
+
+    @property
+    @deprecated(
+        reason="Attribute `original_values` is deprecated and will be removed! Use `segment_outliers` instead.",
+        version="3.0",
+    )
+    def original_values(self) -> Optional[Dict[str, List[pd.Series]]]:
+        """Backward compatibility property."""
+        if self.segment_outliers is not None:
+            return self.segment_outliers.copy()
+        return None
 
     def get_regressors_info(self) -> List[str]:
         """Return the list with regressors created by the transform.
@@ -39,23 +63,6 @@ class OutliersTransform(ReversibleTransform, ABC):
             List with regressors created by the transform.
         """
         return []
-
-    def _save_original_values(self, ts: TSDataset):
-        """
-        Save values to be replaced with NaNs.
-
-        Parameters
-        ----------
-        ts:
-            original TSDataset
-        """
-        if self.outliers_timestamps is None:
-            raise ValueError("Something went wrong during outliers detection stage! Check the transform parameters.")
-        self.original_values = dict()
-        for segment, timestamps in self.outliers_timestamps.items():
-            segment_ts = ts[:, segment, :]
-            segment_values = segment_ts[segment_ts.index.isin(timestamps)].droplevel("segment", axis=1)[self.in_column]
-            self.original_values[segment] = segment_values
 
     def _fit(self, df: pd.DataFrame) -> "OutliersTransform":
         """
@@ -72,8 +79,7 @@ class OutliersTransform(ReversibleTransform, ABC):
             instance with saved outliers
         """
         ts = TSDataset(df, freq=pd.infer_freq(df.index))
-        self.outliers_timestamps = self.detect_outliers(ts)
-        self._save_original_values(ts)
+        self.segment_outliers = self.detect_outliers(ts)
         self._fit_segments = ts.segments
 
         return self
@@ -99,13 +105,18 @@ class OutliersTransform(ReversibleTransform, ABC):
         NotImplementedError:
             If there are segments that weren't present during training.
         """
-        if self.outliers_timestamps is None:
+        if self.segment_outliers is None:
             raise ValueError("Transform is not fitted! Fit the Transform before calling transform method.")
-        segments = df.columns.get_level_values("segment").unique().tolist()
+
+        segments = set(df.columns.get_level_values("segment"))
+        index_set = set(df.index.values)
+
         check_new_segments(transform_segments=segments, fit_segments=self._fit_segments)
-        for segment in segments:
+        for segment in self.segment_outliers:
+            if segment not in segments:
+                continue
             # to locate only present indices
-            segment_outliers_timestamps = df.index.intersection(self.outliers_timestamps[segment])
+            segment_outliers_timestamps = list(index_set.intersection(self.segment_outliers[segment].index.values))
             df.loc[segment_outliers_timestamps, pd.IndexSlice[segment, self.in_column]] = np.NaN
         return df
 
@@ -130,13 +141,20 @@ class OutliersTransform(ReversibleTransform, ABC):
         NotImplementedError:
             If there are segments that weren't present during training.
         """
-        if self.original_values is None or self.outliers_timestamps is None:
+        if self.segment_outliers is None:
             raise ValueError("Transform is not fitted! Fit the Transform before calling inverse_transform method.")
-        segments = df.columns.get_level_values("segment").unique().tolist()
+
+        segments = set(df.columns.get_level_values("segment"))
+        index_set = set(df.index.values)
+
         check_new_segments(transform_segments=segments, fit_segments=self._fit_segments)
-        for segment in segments:
-            segment_ts = df[segment, self.in_column]
-            segment_ts[segment_ts.index.isin(self.outliers_timestamps[segment])] = self.original_values[segment]
+        for segment in self.segment_outliers:
+            if segment not in segments:
+                continue
+
+            segment_outliers_timestamps = list(index_set.intersection(self.segment_outliers[segment].index.values))
+            original_values = self.segment_outliers[segment][segment_outliers_timestamps].values
+            df.loc[segment_outliers_timestamps, pd.IndexSlice[segment, self.in_column]] = original_values
         return df
 
     @abstractmethod