+++ tests/test_analysis/test_outliers/test_confidence_interval_outliers.py
@@ -1,4 +1,5 @@
 import numpy as np
+import pandas as pd
 import pytest
 
 from etna.analysis import get_anomalies_prediction_interval
@@ -77,3 +78,15 @@ def test_get_anomalies_prediction_interval_values(outliers_tsds, model, interval
 )
 def test_get_anomalies_prediction_interval_imbalanced_tsdf(imbalanced_tsdf, model, interval_width, in_column):
     get_anomalies_prediction_interval(imbalanced_tsdf, model=model, interval_width=interval_width, in_column=in_column)
+
+
+@pytest.mark.parametrize("index_only, value_type", ((True, list), (False, pd.Series)))
+def test_get_anomalies_prediction_interval_index_only(outliers_tsds, index_only, value_type):
+    result = get_anomalies_prediction_interval(
+        outliers_tsds, model=ProphetModel, interval_width=0.95, in_column="target", index_only=index_only
+    )
+
+    assert isinstance(result, dict)
+    for key, value in result.items():
+        assert isinstance(key, str)
+        assert isinstance(value, value_type)